+++ ruststep/src/ast/parameter.rs
@@ -1,4 +1,4 @@
-use crate::step::*;
+use crate::ast::*;
 use serde::{
     de::{self, VariantAccess},
     forward_to_deserialize_any, Deserialize,
@@ -12,7 +12,7 @@ use std::fmt;
 ///
 /// ```
 /// use nom::Finish;
-/// use ruststep::{parser::exchange, step::Parameter};
+/// use ruststep::{parser::exchange, ast::Parameter};
 ///
 /// // Real number
 /// let (residual, p) = exchange::parameter("1.0").finish().unwrap();
@@ -39,7 +39,7 @@ use std::fmt;
 ///
 /// ```
 /// use nom::Finish;
-/// use ruststep::{parser::exchange, step::Parameter};
+/// use ruststep::{parser::exchange, ast::Parameter};
 ///
 /// let (residual, p) = exchange::parameter("B((1.0, A((2.0, 3.0))))")
 ///     .finish()
@@ -70,7 +70,7 @@ use std::fmt;
 /// Create a list as `Parameter::List` from `Iterator<Item=Parameter>` or `Iterator<Item=&Parameter>`.
 ///
 /// ```
-/// use ruststep::step::Parameter;
+/// use ruststep::ast::Parameter;
 ///
 /// let p: Parameter = [Parameter::real(1.0), Parameter::real(2.0)]
 ///     .iter()
@@ -88,7 +88,7 @@ use std::fmt;
 ///
 /// ```
 /// use serde::Deserialize;
-/// use ruststep::step::Parameter;
+/// use ruststep::ast::Parameter;
 ///
 /// #[derive(Debug, Deserialize)]
 /// struct A {
@@ -137,7 +137,7 @@ use std::fmt;
 ///
 /// ```
 /// use serde::Deserialize;
-/// use ruststep::{parser::exchange, step::RValue};
+/// use ruststep::{parser::exchange, ast::RValue};
 /// use nom::Finish;
 ///
 /// let (res, p) = exchange::parameter("#11").finish().unwrap();
