+++ ruststep/src/header.rs
@@ -1,4 +1,4 @@
-use crate::{error::Result, step::*};
+use crate::{ast::*, error::Result};
 use serde::Deserialize;
 
 #[derive(Debug, Clone, PartialEq, Deserialize)]
