+++ ruststep/src/parser/exchange/anchor.rs
@@ -1,6 +1,6 @@
 use crate::{
+    ast::*,
     parser::{combinator::*, token::*},
-    step::*,
 };
 use nom::{branch::alt, Parser};
