+++ ruststep/src/parser/exchange/reference.rs
@@ -1,6 +1,6 @@
 use crate::{
+    ast::*,
     parser::{combinator::*, token::*},
-    step::*,
 };
 use nom::Parser;
