+++ ruststep/src/parser/mod.rs
@@ -38,8 +38,8 @@ pub mod exchange;
 pub mod token;
 
 use crate::{
+    ast::*,
     error::{Result, TokenizeFailed},
-    step::*,
 };
 use nom::Finish;
