+++ ruststep/src/tables.rs
@@ -49,7 +49,7 @@
 //!     - `a_iter()` as above example
 //!
 
-use crate::step::*;
+use crate::ast::*;
 use serde::{
     de::{self, IntoDeserializer, VariantAccess},
     Deserialize,
